requirements.txt
requests
Flask
oauthlib
pyOpenSSL
Flask-Login
python-dotenv
supabase